Relocate shared code into src/sources/util/http.rs
